pyproject.toml
[project]
name = "baostock-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "backtesting>=0.6.5",
    "baostock>=0.8.9",
    "gradio>=6.6.0",
    "numpy<2.3.0",
    "pandas<3.0.0",
    "pandas-ta>=0.4.71b0",
    "pyecharts>=2.1.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "tqdm>=4.67.3",
]
